pyproject.toml
[tool.poetry]
name = "curs-work-7"
version = "0.1.0"
description = ""
authors = ["Сергей <sergey.vasilev.98@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.9"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.28.0"
stripe = "^11.3.0"
forex-python = "^1.8"
django-cors-headers = "^4.6.0"
celery = "^5.4.0"
eventlet = "^0.38.2"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
